Resolved msvc C4302 (PVOID to unsigned long conversion)

+++ src/shared/CrashHandler.cpp
@@ -202,7 +202,7 @@ void PrintCrashInformation(PEXCEPTION_POINTERS except)
     echo("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-\n");
     echo("Server has crashed. Reason was:\n");
     echo("   %s at 0x%08X\n", GetExceptionDescription(except->ExceptionRecord->ExceptionCode),
-         (unsigned long)except->ExceptionRecord->ExceptionAddress);
+         except->ExceptionRecord->ExceptionAddress);
 #ifdef REPACK
     echo("%s repack by %s has crashed. Visit %s for support.", REPACK, REPACK_AUTHOR, REPACK_WEBSITE);
 #endif
